rebase patches
the history of this patch is a little convoluted; the patches were first rebased from
v0.9 to v1.2 in a688ac0348fcf9203459996598eb28128530c82d; this patch was then reverted for
publishing v1.0 (see #65), so that the patches could be properly rebased from v0.9 to v1.0.

In the meantime, the patches as of v1.2 applied as-is to v1.1 (see #64), until we had to add
some extra backports for the latter in 33e69eacdfe073c1c55f81933c943a7f89d3691f. To backport
those patches to v1.0, we needed another dependent patch that had landed in v1.1, which is
what got publised as v1.0, and then v1.1 got rebased on top of that.

In effect, what we're doing here is just restoring the state of 33e69eacdfe073c1c55f81933c943a7f89d3691f
on top of the patches as of v1.0, which effectively reapplies a688ac0348fcf9203459996598eb28128530c82d
and drops patches that never ended up being necessary for v1.1